rental create and edit fixed
propertytype dropdown fixed working with rentalVm
car Advance search modified

+++ SO.SilList.Tests/Managers/RentalManagerTest.cs
@@ -50,7 +50,7 @@ namespace SO.SilList.Tests.Managers
         public void insertDeleteTest()
         {
             var vo = new RentalVo();
-            vo.price = 1500.99;
+            vo.price = (decimal)1500.99;
 
 
             var result = rentalManager.insert(vo);
